Fix wand arc bug and turn off motion blur
wand arc was defaulting to thumb wheel position 0 when I would open the
config panel because I never set up the default value in the panel
prefab after adding the ability to have default values for thumb wheels.

HDRP motion blur was on and didn't look good through the glasses.